cleaned up main
